+++ examples/ai-core/src/generate-text/anthropic-reasoning-chatbot.ts
@@ -1,4 +1,4 @@
-import { anthropic } from '@ai-sdk/anthropic';
+import { anthropic, AnthropicProviderOptions } from '@ai-sdk/anthropic';
 import { CoreMessage, generateText } from 'ai';
 import 'dotenv/config';
 import * as readline from 'node:readline/promises';
@@ -25,7 +25,7 @@ async function main() {
       providerOptions: {
         anthropic: {
           thinking: { type: 'enabled', budgetTokens: 12000 },
-        },
+        } satisfies AnthropicProviderOptions,
       },
     });