+++ packages/openai-compatible/src/openai-compatible-provider.ts
@@ -1,5 +1,6 @@
 import {
   EmbeddingModelV1,
+  ImageModelV1,
   LanguageModelV1,
   ProviderV1,
 } from '@ai-sdk/provider';
@@ -10,12 +11,15 @@ import { OpenAICompatibleCompletionLanguageModel } from './openai-compatible-com
 import { OpenAICompatibleCompletionSettings } from './openai-compatible-completion-settings';
 import { OpenAICompatibleEmbeddingModel } from './openai-compatible-embedding-model';
 import { OpenAICompatibleEmbeddingSettings } from './openai-compatible-embedding-settings';
+import { OpenAICompatibleImageSettings } from './openai-compatible-image-settings';
+import { OpenAICompatibleImageModel } from './openai-compatible-image-model';
 
 export interface OpenAICompatibleProvider<
   CHAT_MODEL_IDS extends string = string,
   COMPLETION_MODEL_IDS extends string = string,
   EMBEDDING_MODEL_IDS extends string = string,
-> extends ProviderV1 {
+  IMAGE_MODEL_IDS extends string = string,
+> extends Omit<ProviderV1, 'imageModel'> {
   (
     modelId: CHAT_MODEL_IDS,
     settings?: OpenAICompatibleChatSettings,
@@ -40,6 +44,11 @@ export interface OpenAICompatibleProvider<
     modelId: EMBEDDING_MODEL_IDS,
     settings?: OpenAICompatibleEmbeddingSettings,
   ): EmbeddingModelV1<string>;
+
+  imageModel(
+    modelId: IMAGE_MODEL_IDS,
+    settings?: OpenAICompatibleImageSettings,
+  ): ImageModelV1;
 }
 
 export interface OpenAICompatibleProviderSettings {
@@ -85,12 +94,14 @@ export function createOpenAICompatible<
   CHAT_MODEL_IDS extends string,
   COMPLETION_MODEL_IDS extends string,
   EMBEDDING_MODEL_IDS extends string,
+  IMAGE_MODEL_IDS extends string,
 >(
   options: OpenAICompatibleProviderSettings,
 ): OpenAICompatibleProvider<
   CHAT_MODEL_IDS,
   COMPLETION_MODEL_IDS,
-  EMBEDDING_MODEL_IDS
+  EMBEDDING_MODEL_IDS,
+  IMAGE_MODEL_IDS
 > {
   const baseURL = withoutTrailingSlash(options.baseURL);
   const providerName = options.name;
@@ -154,6 +165,16 @@ export function createOpenAICompatible<
       getCommonModelConfig('embedding'),
     );
 
+  const createImageModel = (
+    modelId: IMAGE_MODEL_IDS,
+    settings: OpenAICompatibleImageSettings = {},
+  ) =>
+    new OpenAICompatibleImageModel(
+      modelId,
+      settings,
+      getCommonModelConfig('image'),
+    );
+
   const provider = (
     modelId: CHAT_MODEL_IDS,
     settings?: OpenAICompatibleChatSettings,
@@ -163,10 +184,12 @@ export function createOpenAICompatible<
   provider.chatModel = createChatModel;
   provider.completionModel = createCompletionModel;
   provider.textEmbeddingModel = createEmbeddingModel;
+  provider.imageModel = createImageModel;
 
   return provider as OpenAICompatibleProvider<
     CHAT_MODEL_IDS,
     COMPLETION_MODEL_IDS,
-    EMBEDDING_MODEL_IDS
+    EMBEDDING_MODEL_IDS,
+    IMAGE_MODEL_IDS
   >;
 }