+++ packages/solid/src/use-object.ts
@@ -72,6 +72,13 @@ export type Experimental_UseObjectOptions<RESULT> = {
    * Additional HTTP headers to be included in the request.
    */
   headers?: Record<string, string> | Headers;
+
+  /**
+   * The credentials mode to be used for the fetch request.
+   * Possible values are: 'omit', 'same-origin', 'include'.
+   * Defaults to 'same-origin'.
+   */
+  credentials?: RequestCredentials;
 };
 
 export type Experimental_UseObjectHelpers<RESULT, INPUT> = {
@@ -163,6 +170,7 @@ function useObject<RESULT, INPUT = any>(
           'Content-Type': 'application/json',
           ...useObjectOptions().headers?.(),
         },
+        credentials: useObjectOptions().credentials?.(),
         signal: abortController.signal,
         body: JSON.stringify(input),
       });