+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,35 @@
 
 ### Patch Changes
 
+- Updated dependencies [625591b]
+- Updated dependencies [6a1506f]
+- Updated dependencies [ea3d998]
+  - ai@4.2.6
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.2.5
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [3d6d96d]
+  - ai@4.2.4
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0b3bf29]
+  - ai@4.2.3
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [f10f0fa]
   - ai@4.2.2